Add tests for user_service.py
