+++ requirements.txt
@@ -3,5 +3,6 @@ waitress==1.3.0
 SQLAlchemy==1.3.5
 #test dependencies
 pytest==4.4.1
+pytest-mock==1.10.4
 coverage==4.5.3
 pytest-cov==2.7.1
